+++ src/main/java/edu/ie3/datamodel/io/factory/EntityFactory.java
@@ -23,6 +23,7 @@ public abstract class EntityFactory<T extends Entity, D extends EntityData>
    *
    * @param allowedClasses exactly the classes that this factory is allowed and able to build
    */
+  @SafeVarargs
   protected EntityFactory(Class<? extends T>... allowedClasses) {
     super(allowedClasses);
   }