+++ CHANGELOG.md
@@ -8,10 +8,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+
 ### Fixed
 
 
 ### Changed
+- Switched trait `WeatherSourceTestHelper` to class and more code clean up [#1657](https://github.com/ie3-institute/PowerSystemDataModel/issues/1657)
 - Removed graphic related code [#1655](https://github.com/ie3-institute/PowerSystemDataModel/issues/1655)